Refactor code agent communication and processing
- Introduced `CodeAgentImplementationRequest` and `CodeAgentImplementationTask` for enhanced client-server communication with `ConnectionId`.
- Updated `CodingAgentEndpoints` to use the new request structure and handle tasks asynchronously via `Channel<CodeAgentImplementationTask>`.
- Added `CodeAgentBackgroundService` for improved background task management.
- Modified `MyCloudEventClient` to send messages to specific clients using connection IDs.
- Updated `McpHttpClient` and `McpService` to accommodate new method signatures.
- Added commented-out code sections for potential future implementations and context.
- Improved overall code structure for clarity, maintainability, and better SignalR integration.
