+++ src/MCP.Host.Contracts/CodeAgentImplementationRequest.cs
@@ -1,0 +1,3 @@
+﻿namespace MCP.Host.Contracts;
+
+public record CodeAgentImplementationRequest(string ConnectionId, string Requirement);
