+++ src/MCP.Host/Services/CodeAgentImplementationTask.cs
@@ -1,0 +1,3 @@
+﻿namespace MCP.Host.Services;
+
+public record CodeAgentImplementationTask(string ConnectionId, string Requirement);
