+++ src/MCP.WebApp/MCP.WebApp.Client/Services/IMcpHttpClient.cs
@@ -5,5 +5,5 @@ public interface IMcpHttpClient
     Task<string> ChatAsync(string message, CancellationToken cancellationToken);
     Task AgentStreamAsync(Guid chatId, string message, Action<string> onChunk, CancellationToken cancellationToken);
     Task RemoveChatAsync(Guid chatId, CancellationToken cancellationToken);
-    Task CodeAgentStreamAsync(Guid chatId, string message, Action<string> onChunk, CancellationToken cancellationToken);
+    Task CodeAgentStreamAsync(Guid chatId, string connectionId, string message, Action<string> onChunk, CancellationToken cancellationToken);
 }
