+++ src/MCP.WebApp/MCP.WebApp.Client/Services/McpHttpClient.cs
@@ -37,11 +37,11 @@ public class McpHttpClient(HttpClient httpClient) : IMcpHttpClient
         }
     }
 
-    public async Task CodeAgentStreamAsync(Guid chatId, string message, Action<string> onChunk, CancellationToken cancellationToken)
+    public async Task CodeAgentStreamAsync(Guid chatId, string connectionId, string message, Action<string> onChunk, CancellationToken cancellationToken)
     {
         using var request = new HttpRequestMessage(HttpMethod.Post, "agent/workflow");
         request.Headers.Add(HeaderNames.ChatIdHeaderName, chatId.ToString());
-        request.Content = JsonContent.Create(new ChatRequest(message));
+        request.Content = JsonContent.Create(new CodeAgentImplementationRequest(connectionId, message));
         httpClient.Timeout = TimeSpan.FromMinutes(10);
         var response = await httpClient.SendAsync(request, HttpCompletionOption.ResponseHeadersRead, cancellationToken);
         response.EnsureSuccessStatusCode();
