+++ src/MCP.WebApp/MCP.WebApp.Client/Services/McpService.cs
@@ -17,8 +17,8 @@ public class McpService(IMcpHttpClient mcpHttpClient) : IMcpService
         await mcpHttpClient.RemoveChatAsync(chatId, cancellationToken);
     }
 
-    public async Task CodeAgentStreamAsync(Guid chatId, string message, Action<string> onChunk, CancellationToken cancellationToken)
+    public async Task CodeAgentStreamAsync(Guid chatId, string connectionId, string message, Action<string> onChunk, CancellationToken cancellationToken)
     {
-        await mcpHttpClient.CodeAgentStreamAsync(chatId, message, onChunk, cancellationToken);
+        await mcpHttpClient.CodeAgentStreamAsync(chatId, connectionId, message, onChunk, cancellationToken);
     }
 }
